Sync: Create typescript environment for Sync Engine
Build a typescript directory for Sync to live in. Do not lint or commit
the built output in `/lib/sync`. Bring across a tsconfig.json that has
some common restrictions and minimizes the built output. Require that
this directory is compiled before we run the js based test suite.

Define a rather nascent structure that simply states that an Integration
must have method by which the urls relevant for the Sync Engine may be
extracted. Create a rather nascent GitHub integration that understands
comments and issues and where there urls should be. Put a rather nascent
set of tests through the GitHub integration.

Change-type: patch
Signed-off-by: Andrew Lucas <andrewl@resin.io>