Review linting of the repo
I enabled the previously broken line-length rule. This triggered one
change, which is a line break in the initiation rituals text.

Set eslint to ignore `*/dist/` folders rather than `sdk/*`, so that it
is more compatible with the existence of other typescript modules and
so that non-built js within `/sdk` is linted.

Change-type: patch
Signed-off-by: Andrew Lucas <andrewl@resin.io>