Sync: Added GitHub integration
Defined that an integration must have methods to authenticate, cherry
pick the ids from a payload into a time-line, convert a single
event into a time-line, fetch a full time-line and translate a time-line
into cards.

Wrote a GitHub integration that obeys the above, using the
`@octokit/rest` SDK.

Made a utility function to make JWT encoding recyclable.

Because the GitHub webhooks tell you the type of payload in
`headers['X-GitHub-Event']`` told the `external-event` to keep hold of
the headers object.  This has meant that the permissions test has had to
be tweaked because the card records `headers.host`

Created a batch of tests that will run an integration through
its paces if provided with suitable resources, and applied these to the
GitHub integration.

Change-type: patch
Signed-off-by: Andrew Lucas <andrewl@resin.io>