Sync: Add GitHub integration
Add a single function module, `action-library/sync/github`, that
asynchronously turns action-library style parameters into a set of
JSON-E insertion instructions. This queries the backend to determine if
the head card is already known before requesting either a full time line
of the issue or making just a short list of changes.

Provide example payloads and a few test functions for the above.

Type the points of contact between an action-library function and the
wider code.

Add the ability for an action-library function to query the context.

Change-type: patch
Signed-off-by: Andrew Lucas <andrewl@resin.io>