worker: Only consider type triggers with filters
This is a small PR to set things up for time-triggered actions. These
type of triggered actions only differ in that they won't have a filter,
so we must update the worker to only consider type triggers that do
contain a filter (i.e. are based on insertion of other cards).

Change-type: patch
Signed-off-by: Juan Cruz Viotti <juan@resin.io>