worker: Don't expose source in .getRequest() if no input card
Time triggered actions don't have an input card as there is no insertion
(but time) that triggers them. We still want to be able to generate
action requests out of those, so we must modify `triggers.getRequest()`
to react to the fact that the input card is now optional.

This basically means that the `source` property will not be accessible
in time-based triggered action templates.

Change-type: patch
Signed-off-by: Juan Cruz Viotti <juan@resin.io>