worker: Fix race condition when waiting for a request execution
When the server receives an action request through its HTTP API, it
enqueues it and starts waiting for the result.

The wait function algorithm looks like this:

- Check if the request has been executed already, if so return it
- Open a stream to wait for it

The race condition can happen if the action request is executed after we
checked if the request has been executed, but before we open the stream.
In this case, the worker will just hang forever waiting for the result.

Change-type: patch
Signed-off-by: Juan Cruz Viotti <juan@resin.io>