core: Add non-production runtime assertions on kernel getters
This asserts system can be used for cases that the server can recover
from, but that shouldn't happen in the first place. The assertions will
only run during testing/development and will NOT run in production.

In this case, if two cards have the same ids or slugs, then the server
will be fine with it by just returning the first one, while throwing
would crash the whole system. We still want to catch these cases on our
test runs.

Change-type: patch
Signed-off-by: Juan Cruz Viotti <juan@balena.io>