core: Require type in .getElementBy(Id|Slug) functions
I found that in all cases where we're using these direct getters, we
always know the type we are expecting in result. Therefore it makes
sense to make it required, simplifying a lot of the logic to traverse
through tables, and to make it very hard to introduce performance
regressions that impact database reads.

Change-type: patch
Signed-off-by: Juan Cruz Viotti <juan@balena.io>