kernel: Use a lockfile when bootstrapping the backend
This change allows integration tests to run in parallel, speeding up
test build times.
The portFinder package is now used to find open ports for the web
server, as the previous approach of waiting for an 'error' event to be
emitted would fail due to the `--abort-on-uncaught-exception` flag.

Change-type: patch
Signed-off-by: Lucian <lucian.buzzo@gmail.com>