circleci: Run unit and integration tests multiple times
This should be a good enough solution to prevent us from inserting flaky
tests or flaky components that make the tests sporadically fail. It is
more computational wasteful, but I think its a fair tradeoff given the
amount of time we're spending dealing with sporadic CI failures (when
debugging them or trying to merge something else).

If it becomes a problem, then we can probably get more Circle CI
parallel jobs in our plan.

Change-type: patch
Signed-off-by: Juan Cruz Viotti <juan@balena.io>