circleci: Retry unit/integration tests on the same job
We added this as a way to prevent flaky tests. The retries were running
in parallel on the workflow, but that means that if a job fails, we can
"retry from failed", and avoid the sporadic issues altogether.

By making them all run in one job, we can't really cheat, as all the
retries need to pass in one go for the build to be green.

Change-type: patch
Signed-off-by: Juan Cruz Viotti <juan@balena.io>