logger: Setup a native LogEntries integration
I refactored the logger module to take a LogEntries token, and if that's
available, it will send the logs directly to LogEntries (preserving
object structure) and NOT output to `stdout`, which should also give us
a small boost on production, as we don't have to stringify huge objects,
etc.

Everything behaves as usual if such token is not set.

Change-type: minor
Signed-off-by: Juan Cruz Viotti <juan@balena.io>