queue: Stop keeping track of the queue size
Turns out the current approach is incredibly expensive on a real-world
database with a lot of data. There are better ways to do it which we can
explore in the future, but for now we can give up on it (we only use it
for runtime metrics purposes) so the application is at least not down.

Change-type: minor
Signed-off-by: Juan Cruz Viotti <juan@balena.io>