sync: Fix duplicated Front messages
When syncing messages, we check that neither the database nor the
sequence we are about to insert mention the mirror id we are dealing
with.

However consider that multiple workers might be trying to create the
same message/whisper. Neither of them reference the mirror id in their
sequences, and they check the database at the same time, so they might
still insert the same thing twice. The slug is the only safe way to
prevent this.

Change-type: patch
Signed-off-by: Juan Cruz Viotti <juan@balena.io>