sync: Support an integration default user actor
This user is set as `INTEGRATION_DEFAULT_USER`, and will be used on an
OAuth request if the current actor is not linked to the OAuth service
but the default user is.

Some of this workflow is unit tested, and we will increase the coverage
after building the first OAuth integration.

Change-type: minor
Signed-off-by: Juan Cruz Viotti <juan@balena.io>