UI: Refactor routing to use react-router
This changes drops the previous approach of prefixing the path parm with
the card type and now prefers card slugs over card ids.
Prefixing by card type was an optimisation added to improve performance
on rethinkdb, but now we use postgres there is no performance
improvement and this oddity is not needed.

Change-type: minor
Signed-off-by: Lucian <lucian.buzzo@gmail.com>