Remove investigating category from support threads
After using this category in production it has become a place where
support threads go to die and is largely unmaintained. We determined
that it was harmful and we need to rethink our strategy for support
threads that need long running investigations. In the meantime we are
removing this category

Change-type: minor
Signed-off-by: Lucian <lucian.buzzo@gmail.com>