worker: Fix create/update event sync race condition
Consider a piece of code that checks whether a card exists in order to
call the `.insertCard()` or `.patchCard()` worker method. If the outcome
is to insert, but the card gets inserted in the mean-time, then by the
time the executor runs, it will check again in order to decide if it
should post a `create` or an `update` event, which means we can end up
with a `create` event that has the payload of an `update` event.

Change-type: patch
Signed-off-by: Juan Cruz Viotti <juan@balena.io>