core: Refactor and simplify permission filter code
Most of these test changes are caused by the fact that we are replacing
a custom "merge" logic set of statements with the more proper
`skhema.merge()` function, which highlighted some semantic issues.

As far as I can tell, these changes to how we use `additionalProperties`
are correct, while the previous ones were ambiguous in some cases.

Change-type: patch
Signed-off-by: Juan Cruz Viotti <juan@balena.io>