action-server: Don't enable queue priority buffer on tick server
Enabling the priority buffer on a queue instance means that the queue
also opens a stream for action requests and attempts to consume from
there before finding the action request on the requests table, as an
optimisation.

We don't need this optimisation on the tick's queue, as we don't consume
incoming action requests like the other types of action servers.

Change-type: patch
Signed-off-by: Juan Cruz Viotti <juan@balena.io>