front: Correctly throw error object from request-promise
According to the docs at https://github.com/request/request-promise, the
error object that the promise rejects with looks like this:

```
cause: err,
error: err,
options: options,
response: response
```

And apparently the actual error can be either on `cause` or on `error`,
but its unclear why they have such distinction or on what cases
`error.error` is a string.

We should probably refactor this to not use `request-promise`, but then
we're ditching Front anyways, so it might not be worth the effort.

Change-type: patch
Signed-off-by: Juan Cruz Viotti <juan@balena.io>