core: Ensure that we can update the markers materialized view
As it is now, we only create the `cards_markers` materialized view if it
doesn't exist, which means that any change to the definition of this
file doesn't make it to a database that had already been initialized by
Jellyfish.

I couldn't find a better way to update an existing materialized view (if
needed) that didn't involve dropping the view and re-creating it from
scratch.

The `EXISTS` parameters should ensure that this is safe given multiple
containers trying to do the same thing.

Change-type: patch
Signed-off-by: Juan Cruz Viotti <juan@balena.io>