front: Don't consider private inboxes
If a conversation starts on a private personal inbox and is then moved
to i.e. support, then the conversation will belong to both inboxes at
the same time.

When looking what inbox a thread belongs to, we shouldn't take into
account private ones, as we don't know what to do with them.

Change-type: patch
Signed-off-by: Juan Cruz Viotti <juan@balena.io>