dev: Check node version after npm install
Added a simple script that will error if the node version is less than the one specified in `engines` in `package.json`. This script is called by the postinstall routine. Also added nvm settings file so you can just call `nvm use` to set it to the right version.

Change-type: minor
Signed-off-by: Graham McCulloch <graham@balena.io>