discourse: Use deletor as post actor when user is deleted
Consider that a user sends a post and then the user is hard deleted from
the database as a response to that post (i.e. spam checks). The webhook
event will contain no information about the actor whatsoever, all the
details will be `null`, the API will respond with 404, etc.

The orphaned post then belongs to the "system" Discourse user, which is
the deletor of the post.

This change will at least allow us to sync these deleted messages as
belonging to the "system" user, rather than completely failing the
action.

Change-type: patch
Signed-off-by: Juan Cruz Viotti <juan@balena.io>