chat-widget: Add dev-only support for Redux Dev Tools
Redux Dev Tools is a Chrome Extension that is very useful for debugging the Redux store. Its already supported by the main Jellyfish app. This commit adds the same change to the chat widget, explicitly providing a name for the chat-widget's Redux store so you can distinguish it from the 'Jellyfish' one in the main app.

Change-type: minor
Signed-off-by: Graham McCulloch <graham@balena.io>