core: Make card table creation not lock if not needed
Looks like `CREATE TABLE IF NOT EXISTS` and `ALTER TABLE <XXX> ADD IF
NOT EXISTS` will take exclusive locks on the table even if they are
no-ops.

After this commit, we approximate `CREATE TABLE IF NOT EXISTS` in a
non-locking way by first fetching the list of existing tables. Of
course, this is a non-atomic operation that will prevent the lock
altogether in the best case, or will end up in the current no-op
blocking situation in the worst case.

I think its also safe to remove the `ALTER TABLE` statements below as
all the environments we care about have been already migrated, and we
need a better migration story anyways.

Change-type: patch
Signed-off-by: Juan Cruz Viotti <juan@balena.io>