Automatically create partial indexes based on type cards
A new `indexed_fields` field is added to type cards, this field's value
is an array of arrays of strings. Each sub-array of strings represents
columns in the database. If the column to be indexed is a jsonb column,
it is represented using a '.' delimiter.
On create or update, if a type card has an `indexed_fields` property,
partial indexes will be created for that type using the columns named.

For example, if the link type card has `indexed_fields` set to:
```
[
	[ 'data.from.id', 'name', 'data.to.id' ]
]
```
A single multi-column partial index will be created, constrained to
cards where the type is "link".

Change-type: minor
Signed-off-by: Lucian <lucian.buzzo@gmail.com>