test: Drop SQL tables after running integration tests
Introducing a unique composite constraint on the slug and version
properties for some reason slows down PostgreSQL a lot (>6x) when we
have a lot of databases, which we do when running integration test (we
create a separate database for each test case).

This commit makes sure we drop the tables from the databases after
running each integration test case, which seems to solve the issue.

Change-type: minor
See: https://github.com/product-os/jellyfish/pull/3621
Signed-off-by: Juan Cruz Viotti <juan@balena.io>