Improve performance of kernel integration tests
Previously, the kernel tests would configure new backend and kernel
instances for every test, allowing the tests to run in a pristine
environment.
This commit makes the bulk of the kernel tests share a backend and kernel
instance, which required a number of refactors to ensure that the tests
still work, even when there are many more cards in the system.
Some tests have remained as they were, as they explicitly test
disconnects, requiring them to have seperate instances.

In addition to these changes, duplicate tests for skip and limit options
have been removed.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>