ui: Convert Jellyfish into a Progressive Web App
Added a manifest and a service worker. Jellyfish can now be 'installed' as a Progressive Web App. All static resources are pre-cached by the service worker using a 'CacheFirst' strategy. When the app is updated in any way, the user is notified via a confirm dialog and the app is updated and the page is refreshed once they accept.

Change-type: minor
Signed-off-by: Graham McCulloch <graham@balena.io>