Quote all the things (in k8s templates)
Previously we were relying on the value to look like a string, which led
to spurious quoting for variables that could have a number-like value
like a port.

Also templeting was used inconsistently by using double brackets
({{...}}) which HTML escapes instead of triple ones that put the value
as-is.

Both issued are fixed by universally quoting all values and converting
everything to triple brackets.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>
Change-Type: patch