Explicitly set which links table columns that values are written to
By explicitly setting which columns alues are written to, we avoid an
error where the insert would fail if the table columns were in
a different order to the values being inserted. This can happen if
a previously created table is altered to include the new version
columns.

Change-type: patch
Signed-off-by: Lucian <lucian.buzzo@gmail.com>