Fix edge case where a subschema of type object could produce invalid SQL
The JSON schema keywords `contains` and `items` can be implemented as
subqueries. Both old and new compilers were not handling this correctly
when the `contains`/`items` schema was an object with a property
restricetd by a keyword that requires JSON-specific text-extraction
operators (`->>`/`#>>`). For example, the `pattern` keyword.

This commit fixes this edge case in the new compiler and adds a test for
this edge case that is skipped by default. This test should be enabled
when the new compiler becomes the default or if the old one is patched.

Change-type: patch
Signed-off-by: Carol Schulze <carol@balena.io>