ui: Ensure search works on all inbox tabs
Needed to be careful to avoid duplications when merging `required` fields - hence the use of `_.mergeWith` with a customizer function. Abstracting out the search filter to a single method will also prove useful when we start searching for `searchable` fields and the search filter has to be modified.

Change-type: patch
Signed-off-by: Graham McCulloch <graham@balena.io>