Refactor link traversal to use main table
This change refactors link traversal to use the main table, taking advantage of partial indexes that have been created. Previously whenever a link was created, modified or deleted, a row on the `links` table was updated. The `links` table was then used to traverse relationships when querying with `$$links`. This was a performance optimization that is not needed now that we have partial indexes on the main table.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>