Reduce emission of type-checking conditionals in the new compiler
Many JSON schema keywords must be implemented with a material
conditional at the SQL level, equivalent to:

```
typeof <field> === <type> ? <condition> : true
```

This is implemented as a boolean expression at the SQL level:

```
NOT <isType> OR <condition>
```

where `<isType>` is the equivalent SQL for the `typeof` check.
Frequently the type is specified in the schema, leading to the following
SQL expression:

```
<isType> AND (NOT <isType> OR <condition>)
```

which siplifies to just `<isType> AND <condition>`. PG does not
recognize this optimization opportunity so this commit applies it at the
compiler level.

Change-type: patch
Signed-off-by: Carol Schulze <carol@balena.io>