core: Always replace card with NULL version fragments if any
Right now we set a default on NULL fragments when reading them from the
database. Attempting to update one of these unversioned cards will
result in a version bump as they obtain the new version.

This change adds logic to update the unversioned card in place for such
scenarios.

Change-type: patch
Signed-off-by: Juan Cruz Viotti <juan@balena.io>