Correctly deal with arrays containing objects in the new compiler
The `contains` and `items` keywords are implemented as subqueries in the
general case. The `SqlQuery` instance handling this subquery has to know
whether it is checking the contents of a PG array, or a JSON array so it
can generate the correct field access. Handling of this flag was faulty
and an incorrect field access was being generated.

This commit also includes a test case that is skipped by default and
should be enabled when the new compiler becomes the default.

Chage-type: patch
Signed-off-by: Carol Schulze <carol@balena.io>