Improve favicon rendering
Firefox does not work with a source svg file unless the width and height are specified. So we use separate png files of the correct sizes instead. We also now define multiple sizes of favicons (at the moment 16 and 32). The font is also now a bit bolder so that it still renders well on Firefox on Windows.

Change-type: patch
Signed-off-by: Graham McCulloch <graham@balena.io>