Only reopen a support thread linked to an issue on the "closed" event
The triggered action that re-opens a support thread when an issue it's
attached to is closed has been changed so that it will only trigger when
an update event that closes the issue is created.
Previously the trigger would run on every worker job that matched the
state where an issue and a support thread are linked and they are both
closed. This had the side effect of triggering the action when you link
to an issue that is already closed, or if you edit a field of a closed
issue.

Connects-to: #4039
Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>