Use raw discourse posts instead of the "cooked" HTML versions.
Previously we were working with the HTML representations of messages
from discourse, instead of the "raw" (usually markdown formatted)
messages. This fixes an issue where after posting a message from
Jellyfish, the message would be updated with the HTML representation
generated by discourse. It also means that #summary detection should
work correctly for message that originate in discourse.

Connects-to: #3728
Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>