Remove "slices" definition from support issues
The current implementation of the "slices" functionality impedes the
usefulness of the support issues, as it preselects high severity issues,
which is not what is wanted by support agents. The "slices" concept most
likely needs to be rethought, but for the time being, removing the
annotation from the support issue card type will provide a good
quality-of-life improvement for support agents looking through support
issues.

Change-type: minor
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>