Move ui up to top level of redux state
Flattening out the redux state structure will make it much simpler to manage what parts of the state are persisted by redux-persist (using whitelist/blacklist).

Note - this will cause some persisted state to be lost when this version is deployed, notably the current expanded items in the sidebar and whether the chat widget is open or not. But this is probably not worth creating a migration for!

Change-type: patch
Signed-off-by: Graham McCulloch <graham@balena.io>