ui: Persist redux state using redux-persist
We still use localForage as the storage engine. Persist and 're-hydration' logic is now greatly simplified and handled automatically by the redux-persist config and the `persistStore` callback ( `onHydrated`).

The `setCoreState` action creator is no longer used/needed so it has been removed.

The `PersistGate` component prevents the child components from rendering until the redux state has been re-hydrated. In practice this delay is so short as to be undetectable.

Note - when this change is deployed, all users will have to login again as location of the cached session token will have changed.

Change-type: patch
Signed-off-by: Graham McCulloch <graham@balena.io>