test: wait for thread sync whisper after creating support-thread
This should fix the race condition seen where sometimes the first 'test' message in a thread is saved before the 'sync whisper' that is created by the asynchronous triggered action. This makes it hard to determine which event will be the 'last' in the list of events. Hopefully by waiting for the sync whisper before posting any test message to the thread, we ensure the sync message will always be the first one in the thread.

Change-type: patch
Signed-off-by: Graham McCulloch <graham@balena.io>