ui: Ensure full message text is always sent
The `newMessage` state is set on a 100ms debounced callback. If the user enters the 'Send Command' within 100ms of typing the last characters of their message, those characters were not getting sent. This change uses the current value of the DOM input element rather than the value cached in the `Timeline`'s state.

Closes #3896

Change-type: patch
Signed-off-by: Graham McCulloch <graham@balena.io>