ui: Add support for linking to a message within a thread
New `Copy link to this message` context menu item. The event's ID is added to the URL as a query param.

When the timeline is loaded, if there is an `event` query param in the URL, we scroll to that event. The focused event pulses the CSS outline property 3 times when it is displayed.

Change-type: minor
Signed-off-by: Graham McCulloch <graham@balena.io>