Refactor flow actions and deprecate LinksProvider component
The LinksProvider component was designed to be used with a single SlideInFlowPanel and a single withLinks child component. Using it on a view wil multiple withLinks child components was not possible. This together with the introduction of optional link queries we have no reason to query for card links seperately. We can remove the LinksProvider logic and streamline our SlideInFlowPanel.

This commit also refactors the object stucture to be referenced on the channel target instead of just the card id. With this we can use a single SlideInFlowPanel component with multiple components controlling it's state.

Change-type: major
Signed-off-by: Stef Kors <stef@balena.io>