Validate registry authentication requests according to docker token auth spec
The `/registry` folder contains a docker registry setup that can be used for development purposes. I'm leaving it here for now as it is useful to have, but we need to have a discussion about where this should live and where the registry setup should be pulled from.

The `api/v2/registry` endpoint is built and tested to comply with the [docker registry v2 token auth spec](https://docs.docker.com/registry/spec/auth/token/) so will work with any docker registry that uses the token auth spec, provided that the Jellyfish API and the registry share the same certificate key environment variable.

Change-type: minor
Signed-off-by: Lucian Buzzo <lucia.buzzo@gmail.com>