Prevent lens selection component collapsing
This change ensures that the lens selection component does not collapse it's height if there are no lenses passed to it. This prevents the UI from jumping around.

The component is also exported as a named export now and the spacing is passed in as a prop now.

Change-type: patch
Signed-off-by: Graham McCulloch <graham@balena.io>