Fix link on SingleCard snippet lens
The link should append the target channel to the current channel (i.e. replacing any existing content to the right of the current channel).

This change introduces a React context which supplies the channel value. This is useful for any components that need to know what channel they are being displayed in.

Change-type: minor
Signed-off-by: Graham McCulloch <graham@balena.io>