Automatically select list lens options
Rather than getting the list of lenses from the hard-coded values in the type card, lenses will now be fetched based on their validity and scoring. All list lenses that are valid for the given data are grouped by lens 'icon' and the top-scoring valid lens for each icon is returned.

Change-type: minor
Signed-off-by: Graham McCulloch <graham@balena.io>