Save previously selected slice to user's profile.
The most recently selected slice will be saved to a user's profile in the same way that the most recently selected lens currently is. This way if you always view a particular view with a slice other than the default first one, this will be preserved and you won't always have to re-select your preferred slice!

Change-type: minor
Signed-off-by: Graham McCulloch <graham@balena.io>