HOTFIX: In views, set the default number of items per page to 100
The lens queryoptions should be overriding the default query options on
the view, but this is not happening on initial load since the active
lens is still unknown while the data is loading.
This hotfix sets the default limit value to 100 as a temporary measure
whilst a root-cause fix is prepared.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>