Use mock data in a View when selecting a lens and canonical data hasn't loaded
Previously, when loading a view, the lenses could not be selected until
the view results are known, as the view results are used to find the
lenses (using their "filters"). This creates problems where lenses can't
be rendered ahead of time, which can be a problem if the lens provides
some kind of query override, and also leads to a jumpy UI as the
available lenses disappear and reappear.
This change uses json-schema-faker to create a mock data object based on
the view query schema, which is then used to select a set of lenses.
This mocking only happens if there are no results available from the
view query.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>