Add omni-search view
Omni-search is a view that lets you search for cards across all searchable card types. It can be accessed via a new searchbox in the sidebar (or from the URL `/search`). A new omni-search lens wraps the normal view lens, constructing and passing in a filter based on the available searchable card types.

Change-type: minor
Signed-off-by: Graham McCulloch <graham@balena.io>