Only merge at most 1 type schema in table lens and sort button
This is a heuristic to avoid errors thrown by the skhema module when
merging incompatible schemas.
As a todo item, we need to improve safety of skhema.merge so that it doesn't
throw if the skhemas can't be merged. That way we can merge all the type schemas
instead of just the first one.

I'm happy to make this change as we currently only support 1 additional
type schema in these components as it is.

Change-type: minor
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>