Restrict view filters to schema of first tail type
This is a heuristic to avoid errors thrown by the skhema module when
merging incompatible schemas.
As a todo item, we need to improve safety of skhema.merge so that it doesn't
throw if the skhemas can't be merged. That way we can merge all the type schemas
instead of just the first one.

Change-type: patch
Signed-off-by: Graham McCulloch <graham@balena.io>