getLinks should require target type
The getLinks action creator should require the target type as well as the verb - otherwise we can fetch cards of the wrong type that also has a link with the same verb.

Note - getLinks is currently just called from Segment and nowhere else so this is quite a constrained bug.

ASAP we should fix the client-sdk's `getWithLinks` method so that it requires the target type(s) to be provided as well. This will of course be a major bump as the API will change.

Change-type: patch
Signed-off-by: Graham McCulloch <graham@balena.io>