Update e2e tests for MyUser lens improvements
Note - I've removed the test for setting the send command to ctrl+enter as it doesn't add much value over the other test for setting the send command to enter (particularly as you can't test the result - see commented out code).

No e2e test has been added for the disableNotificationSound setting as this is tested in a unit test.

Change-type: patch
Signed-off-by: Graham McCulloch <graham@balena.io>