Expose all relationships in tabs
Abstract this functionality into a reusable RelationshipsTab component.

Also refactored the CreateLens and EditLens components to always display the submit and cancel buttons at the bottom and make the form itself scrollable. This should improve the UX as the user can always quickly press cancel or submit and its clear (from the submit disabled state) whether the form data is valid.

Change-type: minor
Signed-off-by: Graham McCulloch <graham@balena.io>