Add a link the the Jellyfish "living spec" in README
Many parts of Jellyfish are still under development, and this document aims to
create a source of truth for how the platform *should* work, even if it doesn't
represent the current functionality. As we develop the design further
and make decisions on functionality, this document will be updated.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>