Run a reduced set of tests against prior Postgres dumps
The intention of running the tests against a previous dump of the DB is
to rule out a category of errors where basic operations would fail when
bootstrapping the stack with a pre-existing database. As such we only
need to run a limited set of smoke tests to verify this.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>