Add loop drop-down to sidebar
When a loop is selected, the loop is used to create a global query mask for the SDK that will affect all subsequent SDK calls. The app is then re-bootstrapped (using this global query mask).

The active loop is also saved to the user's profile so that it persists across sessions.

Note - for now, contracts with no loop specified will always be returned regardless of what the global loop filter is. This allows us to slowly migrate contracts in the appropriate loops. Once all contracts have been migrated to a specific  loop, we can modify the global query mask to filter out contracts with no loop specified when filtering by a particular loop.

Change-type: minor
Signed-off-by: Graham McCulloch <graham@balena.io>