Upgrade to v8 worker
This update change the "trigger" input for the worker to be a plain
triggered-action contract, instead of a parsed intermediary object.
The only change we need to make to support this new worker version is to
no longer transform triggered action contracts before passing them to
the worker.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>