Add a simple UI for check-runs
This commit adds a basic UI for check runs that shows all of the
contracts created by transformers from the commit the check run is
attached to. The graph is generated using mermaidJS, so doesn't add any
additional dependencies to the system. In the future I'd like to develop
this concept into a generic graph visualisation features, however there
are behavioural issues that need to be resolved, such as which links you
expand/display and to what depth.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>