Merge action-server and server services into a single instance
Since both services have the same dependencies, running the core, worker
and queue modules as well as loading plugins, the only realy difference
is that the server can receive external http traffic. Unifying the two
services is a no-brainer and brings us closer to being able to create
a unified jellyfish system, with less indirection and more simplicity.

Change-type: major
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>