Make security test for user create events more targeted.
The previous version of this test did not query for the create contract
directly, which exposed it to subtle bugs, such as `.getWithTimeline()`
returning null. This change makes the test more targeted and explicitly
tests that *other* users cannot read the create event, which is more
close to the intention of the test.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>