Remove MermaidEditor component
The mermaid editor bundles the Monaco editor, which is a very large
package and adds considerable weight to the UI for what is a relatively
small feature that is almost never used. I suggest removing Monaco and
the specialised mermaid editor for the time being and re-introducing
a generic code editing function at some point in the future, possible
lazy loaded or added as a seperate service, dedicated to editing
functionality.

Change-type: major
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>