Improve reliability of livechat e2e tests
This change fixes an issue where types inserted after the initial worker
bootstrap would not be updated correctly in the worker cache. As
a result of this, notification generation could fail if the notification
contract type was inserted after the worker bootstrap phase, resulting
in notifications not being generated.
This change also improves the retry logic when checking for notifications.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>