Use generic contract UI features for support threads
This replaces much of the hardcoding on the support thread lens with the
generic capabilities created for rendering contracts. The most notable
changes is that the auto-generated "Links" tab is now available on
support threads and the timeline is in a seperate tab to the thread
info.
This change also displays the created and updated timestamps on the
`info` tab, making it more discoverable for all contracts.

Change-type: minor
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>