Remove guided teardown and handover panels
The guided flow functionality is tied to work that has stalled and is
will not be progressing at any time in the near future. When we are
ready to tackle handover and teardown workflows again we can revive the
components and functionality and make a targeted team effort to
implement a cohesive process around this functionality.
Currently there isn't a clear process for how to use these features, and
they may no longer be fit for purpose. Retaining them is simply adding
overhead and maintenance burden to the team

This change also removes the ability to assign a contract to someone
else. I strongly feel that this emphasises top-down delegation, when we
should be emphasising concensual ownership. If someone takes ownership
of a contract, they should do it themselves.

Change-type: major
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>